IN-376 feat: add post view for post in feed
